requirements.txt
asgiref==3.5.2
Django==3.2.7
django-3-jet==1.0.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
Unidecode==1.3.4
